Update datasets.py
I was able to work around an error when connecting to an rtsp-stream server and had a division by zero error at this line. I claim, that at the beginning of stream decoding the fps information is wrong (zero). At this point the sleep value shall be a valid constant. I asume 0.2 is ok?